+++ module/src/main/java/com/annimon/tgbotsmodule/api/methods/send/SendVenueMethod.java
@@ -76,23 +76,23 @@ public class SendVenueMethod implements
     }
 
     @Override
-    public Float getLatitude() {
+    public Double getLatitude() {
         return method.getLatitude();
     }
 
     @Override
-    public SendVenueMethod setLatitude(@NotNull Float latitude) {
+    public SendVenueMethod setLatitude(@NotNull Double latitude) {
         method.setLatitude(latitude);
         return this;
     }
 
     @Override
-    public Float getLongitude() {
+    public Double getLongitude() {
         return method.getLongitude();
     }
 
     @Override
-    public SendVenueMethod setLongitude(@NotNull Float longitude) {
+    public SendVenueMethod setLongitude(@NotNull Double longitude) {
         method.setLongitude(longitude);
         return this;
     }
